Split tests into individual files to parallelize them
